refactor(student/dashboard): split performance widgets into separate components
